Demo/requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
absl-py=1.4.0=pypi_0
anyio=3.6.2=pypi_0
argon2-cffi=21.3.0=pypi_0
argon2-cffi-bindings=21.2.0=pypi_0
astor=0.8.1=pypi_0
attrs=22.2.0=pypi_0
backcall=0.2.0=pypi_0
beautifulsoup4=4.11.2=pypi_0
bleach=6.0.0=pypi_0
ca-certificates=2023.01.10=h06a4308_0
certifi=2022.12.7=py37h06a4308_0
cffi=1.15.1=pypi_0
cycler=0.11.0=pypi_0
debugpy=1.6.6=pypi_0
decorator=5.1.1=pypi_0
defusedxml=0.7.1=pypi_0
entrypoints=0.4=pypi_0
fastjsonschema=2.16.2=pypi_0
fonttools=4.38.0=pypi_0
gast=0.2.2=pypi_0
google-pasta=0.2.0=pypi_0
grpcio=1.52.0=pypi_0
h5py=3.8.0=pypi_0
idna=3.4=pypi_0
importlib-metadata=6.0.0=pypi_0
importlib-resources=5.10.2=pypi_0
ipykernel=6.16.2=pypi_0
ipython=7.34.0=pypi_0
ipython-genutils=0.2.0=pypi_0
ipywidgets=8.0.4=pypi_0
jedi=0.18.2=pypi_0
jinja2=3.1.2=pypi_0
joblib=1.2.0=pypi_0
jsonschema=4.17.3=pypi_0
jupyter-client=7.4.9=pypi_0
jupyter-core=4.12.0=pypi_0
jupyter-server=1.23.5=pypi_0
jupyterlab-pygments=0.2.2=pypi_0
jupyterlab-widgets=3.0.5=pypi_0
keras-applications=1.0.8=pypi_0
keras-preprocessing=1.1.2=pypi_0
kiwisolver=1.4.4=pypi_0
ld_impl_linux-64=2.38=h1181459_1
libffi=3.4.2=h6a678d5_6
libgcc-ng=11.2.0=h1234567_1
libgomp=11.2.0=h1234567_1
libstdcxx-ng=11.2.0=h1234567_1
markdown=3.4.1=pypi_0
markupsafe=2.1.2=pypi_0
matplotlib=3.5.3=pypi_0
matplotlib-inline=0.1.6=pypi_0
mistune=2.0.4=pypi_0
nbclassic=0.5.1=pypi_0
nbclient=0.7.2=pypi_0
nbconvert=7.2.9=pypi_0
nbformat=5.7.3=pypi_0
ncurses=6.4=h6a678d5_0
nest-asyncio=1.5.6=pypi_0
notebook=6.5.2=pypi_0
notebook-shim=0.2.2=pypi_0
numpy=1.21.6=pypi_0
open3d-python=0.7.0.0=pypi_0
opencv-python=4.7.0.68=pypi_0
openssl=1.1.1s=h7f8727e_0
opt-einsum=3.3.0=pypi_0
packaging=23.0=pypi_0
pandocfilters=1.5.0=pypi_0
parso=0.8.3=pypi_0
pexpect=4.8.0=pypi_0
pickleshare=0.7.5=pypi_0
pillow=9.4.0=pypi_0
pip=22.3.1=py37h06a4308_0
pkgutil-resolve-name=1.3.10=pypi_0
prometheus-client=0.16.0=pypi_0
prompt-toolkit=3.0.36=pypi_0
protobuf=3.19.0=pypi_0
psutil=5.9.4=pypi_0
ptyprocess=0.7.0=pypi_0
pycparser=2.21=pypi_0
pygments=2.14.0=pypi_0
pyparsing=3.0.9=pypi_0
pyrsistent=0.19.3=pypi_0
python=3.7.16=h7a1cb2a_0
python-dateutil=2.8.2=pypi_0
pyzmq=25.0.0=pypi_0
readline=8.2=h5eee18b_0
scikit-learn=1.0.2=pypi_0
scipy=1.7.3=pypi_0
send2trash=1.8.0=pypi_0
setuptools=65.6.3=py37h06a4308_0
shapely=2.0.1=pypi_0
six=1.16.0=pypi_0
sniffio=1.3.0=pypi_0
soupsieve=2.3.2.post1=pypi_0
sqlite=3.40.1=h5082296_0
tensorboard=1.15.0=pypi_0
tensorflow-estimator=1.15.1=pypi_0
tensorflow-gpu=1.15.0=pypi_0
termcolor=2.2.0=pypi_0
terminado=0.17.1=pypi_0
threadpoolctl=3.1.0=pypi_0
tinycss2=1.2.1=pypi_0
tk=8.6.12=h1ccaba5_0
tornado=6.2=pypi_0
tqdm=4.64.1=pypi_0
traitlets=5.9.0=pypi_0
typing-extensions=4.4.0=pypi_0
wcwidth=0.2.6=pypi_0
webencodings=0.5.1=pypi_0
websocket-client=1.5.1=pypi_0
werkzeug=2.2.2=pypi_0
wheel=0.37.1=pyhd3eb1b0_0
widgetsnbextension=4.0.5=pypi_0
wrapt=1.14.1=pypi_0
xz=5.2.10=h5eee18b_1
zipp=3.12.1=pypi_0
zlib=1.2.13=h5eee18b_0
